Script/requirements.txt
# Core dependencies
pyspark==3.5.0

# Data processing
pandas==2.0.3
numpy==1.24.3

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Configuration
PyYAML==6.0.1

# Performance monitoring
psutil==5.9.5
prometheus-client==0.17.1

# File format support
openpyxl==3.1.2
pyarrow==12.0.1
fastparquet==2023.7.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
